Version bump for phantomjs changes
I decided on a patch version bump because these changes shouldn't break anything and they don't add any new features. The changes fix a bug where the tool hangs due to the user having a newer version of phantomjs in their path than what is compatible with the tool. The changes fix the issue by bundling its own version of phantomjs as a dependency instead of relying on the version of phantomjs installed in the developer's path.